domain,metrics: move domain metrics, add privilege load counter
